Build bcc with libbpf package
Now when we have libbpf packaged we want it to be used to build bcc,
instead of libbpf Github repo, which is downloaded during the build.

The bcc still produces libbcc.so/libbcc.a with libbpf code inside,
so nothing changes there.. apart from that they are built from libbpf
package now instead of libbpf Github repo

The addition is libbcc-no-libbpf.so library, which will be used in
applications like bpftrace which needs to link libbpf.so by itself
for other features.

Signed-off-by: Jiri Olsa <jolsa@redhat.com>